A lot of styling changes + unauthorized page redirect

+++ front-end/services/MovieService.ts
@@ -1,19 +1,35 @@
 import { Movie } from "@types";
 
+
+// Service methods
 const getAllMovies = () => {
+    const token = JSON.parse(sessionStorage.getItem("loggedInUser") as string).token;
+
+    if (!token) {
+        throw new Error("No token found");
+    }
+
     return fetch(process.env.NEXT_PUBLIC_API_URL + "/media/movies", {
         method: "GET",
         headers: {
             "Content-Type": "application/json",
+            Authorization: `Bearer ${token}`,
         }
 })
 };
 
 const createMovie = async (newMovie: Movie) => {
+    const token = JSON.parse(sessionStorage.getItem("loggedInUser") as string).token;
+
+    if (!token) {
+        throw new Error("No token found");
+    }
+
     const response = await fetch(process.env.NEXT_PUBLIC_API_URL + "/media", {
         method: "POST",
         headers: {
             "Content-Type": "application/json",
+            Authorization: `Bearer ${token}`,
         },
         body: JSON.stringify(newMovie)
 });
@@ -24,10 +40,17 @@ return response.json();
 };
 
 const deleteMovie = async (id: number) => {
+    const token = JSON.parse(sessionStorage.getItem("loggedInUser") as string).token;
+
+    if (!token) {
+        throw new Error("No token found");
+    }
+    
     const response = await fetch(process.env.NEXT_PUBLIC_API_URL + `/media/${id}`, {
         method: "DELETE",
         headers: {
             "Content-Type": "application/json",
+            Authorization: `Bearer ${token}`,
         }
 });
 if (!response.ok) {
